chore: linter changes
