+++ src/utils.js
@@ -358,29 +358,3 @@ export const getPushNotificationSettings = (pushNotification) => {
     showAmountEnabled
   };
 };
-
-/**
- * Get registered tokens from the wallet instance.
- * @param {HathorWallet} wallet
- * @param {boolean} excludeHTR If we should exclude the HTR token.
- * @returns {string[]}
- */
-export async function getRegisteredTokens(wallet, excludeHTR = false) {
-  const htrUid = hathorLib.constants.HATHOR_TOKEN_CONFIG.uid;
-  const tokens = [];
-
-  // redux-saga generator magic does not work well with the "for await..of" syntax
-  // The asyncGenerator is not recognized as an iterable and it throws an exception
-  // So we must iterate manually, awaiting each "next" call
-  const iterator = wallet.storage.getRegisteredTokens();
-  let next = await iterator.next();
-  while(!next.done) {
-    const token = next.value;
-    if ((!excludeHTR) || token.uid !== htrUid) {
-      tokens.push(token.uid);
-    }
-    next = await iterator.next();
-  }
-
-  return tokens;
-}
